// packages/js/data/src/payment-gateway-suggestions/action-types.ts
export enum ACTION_TYPES {
	GET_PAYMENT_GATEWAY_SUGGESTIONS_REQUEST = 'GET_PAYMENT_GATEWAY_SUGGESTIONS_REQUEST',
	GET_PAYMENT_GATEWAY_SUGGESTIONS_SUCCESS = 'GET_PAYMENT_GATEWAY_SUGGESTIONS_SUCCESS',
	GET_PAYMENT_GATEWAY_SUGGESTIONS_ERROR = 'GET_PAYMENT_GATEWAY_SUGGESTIONS_ERROR',
	GET_OFFLINE_PAYMENT_GATEWAYS_REQUEST = 'GET_OFFLINE_PAYMENT_GATEWAYS_REQUEST',
	GET_OFFLINE_PAYMENT_GATEWAYS_SUCCESS = 'GET_OFFLINE_PAYMENT_GATEWAYS_SUCCESS',
	GET_OFFLINE_PAYMENT_GATEWAYS_ERROR = 'GET_OFFLINE_PAYMENT_GATEWAYS_ERROR',
}